+++ basenode.h
@@ -47,22 +47,22 @@ struct base_node : public htab_node_t<node_t>, public keynode_t<uint64_t>, publi
          //    data : flag (u_char), string (string_t)
          //    value: hash(string) (uint64_t)
          //
-         u_int s_data_size(void) const;
+         size_t s_data_size(void) const;
 
-         u_int s_pack_data(void *buffer, u_int bufsize) const;
-         u_int s_unpack_data(const void *buffer, u_int bufsize);
+         size_t s_pack_data(void *buffer, size_t bufsize) const;
+         size_t s_unpack_data(const void *buffer, size_t bufsize);
 
          uint64_t s_hash_value(void) const;
 
-         int64_t s_compare_value(const void *buffer, u_int bufsize) const;
+         int64_t s_compare_value(const void *buffer, size_t bufsize) const;
 
-         static u_int s_data_size(const void *buffer);
+         static size_t s_data_size(const void *buffer);
 
-         static const void *s_field_value(const void *buffer, u_int bufsize, u_int& datasize);
+         static const void *s_field_value(const void *buffer, size_t bufsize, size_t& datasize);
 
          static int64_t s_compare_value_hash(const void *buf1, const void *buf2);
 
-         static bool s_is_group(const void *buffer, u_int bufsize);
+         static bool s_is_group(const void *buffer, size_t bufsize);
 };
 
 #endif // __BASENODE_H
