+++ inode.cpp
@@ -43,7 +43,7 @@ inode_t::inode_t(const char *ident) : base_node<inode_t>(ident)
 // serialization
 //
 
-u_int inode_t::s_data_size(void) const
+size_t inode_t::s_data_size(void) const
 {
    return base_node<inode_t>::s_data_size() + 
             sizeof(uint64_t) * 3 +    // count, files, visit
@@ -53,9 +53,9 @@ u_int inode_t::s_data_size(void) const
             sizeof(uint64_t);       // xfer
 }
 
-u_int inode_t::s_pack_data(void *buffer, u_int bufsize) const
+size_t inode_t::s_pack_data(void *buffer, size_t bufsize) const
 {
-   u_int datasize, basesize;
+   size_t datasize, basesize;
    void *ptr;
 
    basesize = base_node<inode_t>::s_data_size();
@@ -81,10 +81,10 @@ u_int inode_t::s_pack_data(void *buffer, u_int bufsize) const
    return datasize;
 }
 
-u_int inode_t::s_unpack_data(const void *buffer, u_int bufsize, s_unpack_cb_t upcb, void *arg)
+size_t inode_t::s_unpack_data(const void *buffer, size_t bufsize, s_unpack_cb_t upcb, void *arg)
 {
    u_short version;
-   u_int datasize, basesize;
+   size_t datasize, basesize;
    const void *ptr;
 
    basesize = base_node<inode_t>::s_data_size(buffer);
@@ -128,7 +128,7 @@ u_int inode_t::s_unpack_data(const void *buffer, u_int bufsize, s_unpack_cb_t up
    return datasize;
 }
 
-u_int inode_t::s_data_size(const void *buffer)
+size_t inode_t::s_data_size(const void *buffer)
 {
    u_short version = s_node_ver(buffer);
    size_t datasize = base_node<inode_t>::s_data_size(buffer) + 
@@ -149,7 +149,7 @@ u_int inode_t::s_data_size(const void *buffer)
             sizeof(double) * 2;           // avgsize, maxsize
 }
 
-const void *inode_t::s_field_value_hash(const void *buffer, u_int bufsize, u_int& datasize)
+const void *inode_t::s_field_value_hash(const void *buffer, size_t bufsize, size_t& datasize)
 {
    u_short version = s_node_ver(buffer);
    size_t offset = base_node<inode_t>::s_data_size(buffer) + 
@@ -165,7 +165,7 @@ const void *inode_t::s_field_value_hash(const void *buffer, u_int bufsize, u_int
    return (u_char*) buffer + offset;
 }
 
-const void *inode_t::s_field_hits(const void *buffer, u_int bufsize, u_int& datasize)
+const void *inode_t::s_field_hits(const void *buffer, size_t bufsize, size_t& datasize)
 {
    datasize = sizeof(uint64_t);
    return (u_char*) buffer + base_node<inode_t>::s_data_size(buffer);
