+++ rcnode.h
@@ -33,21 +33,21 @@ typedef struct rcnode_t : public base_node<rcnode_t> {
          //
          // serialization
          //
-         u_int s_data_size(void) const;
-         u_int s_pack_data(void *buffer, u_int bufsize) const;
-         u_int s_unpack_data(const void *buffer, u_int bufsize, s_unpack_cb_t upcb, void *arg);
+         size_t s_data_size(void) const;
+         size_t s_pack_data(void *buffer, size_t bufsize) const;
+         size_t s_unpack_data(const void *buffer, size_t bufsize, s_unpack_cb_t upcb, void *arg);
 
          uint64_t s_hash_value(void) const;
 
-         int64_t s_compare_value(const void *buffer, u_int bufsize) const;
+         int64_t s_compare_value(const void *buffer, size_t bufsize) const;
 
-         static u_int s_data_size(const void *buffer);
+         static size_t s_data_size(const void *buffer);
 
-         static const void *s_field_value_hash(const void *buffer, u_int bufsize, u_int& datasize);
-         static const void *s_field_hits(const void *buffer, u_int bufsize, u_int& datasize);
-         static const void *s_field_value_mp_url(const void *buffer, u_int bufsize, u_int& datasize);
-         static const void *s_field_value_mp_method(const void *buffer, u_int bufsize, u_int& datasize);
-         static const void *s_field_value_mp_respcode(const void *buffer, u_int bufsize, u_int& datasize);
+         static const void *s_field_value_hash(const void *buffer, size_t bufsize, size_t& datasize);
+         static const void *s_field_hits(const void *buffer, size_t bufsize, size_t& datasize);
+         static const void *s_field_value_mp_url(const void *buffer, size_t bufsize, size_t& datasize);
+         static const void *s_field_value_mp_method(const void *buffer, size_t bufsize, size_t& datasize);
+         static const void *s_field_value_mp_respcode(const void *buffer, size_t bufsize, size_t& datasize);
 
          static int64_t s_mp_compare_value(const void *buf1, const void *buf2, u_int partid, bool& lastpart);
          static int64_t s_compare_hits(const void *buf1, const void *buf2);
